# statistics: BCE for BinarySearchRemoveVal